pyproject.toml
[project]
name = "pet-alert"
version = "1.1.1"
description = "A web application that allows people to search for missing animals and report those found"
readme = "README.md"
requires-python = ">=3.13.2"
authors = ["Ilya Malashenko (github: melax08)"]
dependencies = [
    "aiogram>=3.19.0",
    "celery>=5.4.0",
    "django>=5.1.7",
    "django-celery-results>=2.5.1",
    "django-cors-headers>=4.7.0",
    "django-csp>=3.8",
    "django-filter>=25.1",
    "django-http-referrer-policy>=1.1.1",
    "django-permissions-policy>=4.25.0",
    "django-phonenumber-field>=8.0.0",
    "django-polymorphic>=4.1.0",
    "django-recaptcha>=4.0.0",
    "django-redis>=5.4.0",
    "django-registration>=5.1.0",
    "django-split-settings>=1.3.2",
    "django-structlog>=9.0.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.1",
    "gunicorn>=23.0.0",
    "phonenumberslite>=9.0.1",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "redis>=5.2.1",
    "sentry-sdk[django]>=2.23.1",
    "setuptools>=77.0.1",
    "sorl-thumbnail>=12.11.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.0.1",
    "django-extra-checks>=0.16.1",
    "django-querycount>=0.8.3",
    "django-test-migrations>=1.4.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "ruff>=0.11.0",
]

[tool.ruff]
indent-width = 4
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM"  # flake8-simplify
]

[tool.ruff.format]
line-ending = "lf"
